+++ src/data/initGameProgress.ts
@@ -140,8 +140,10 @@ const initGameProgress: GameProgress = {
   },
   NormalBattleState: {
     curComponent: "VS",
-    isLeftPlayerCurAttack: false,
-    isLeftPlayerCurSelect: false,
+    // isLeftPlayerInitAttack: false,
+    // isLeftPlayerCurAttack: false,
+    // isLeftPlayerCurSelect: false,
+    // selectedBattleButtons: [],
     battleRound: 1,
     SwapCardState: {
       isCardSwitch: false,
@@ -150,6 +152,7 @@ const initGameProgress: GameProgress = {
     SelectGamePadButtonState: {
       leftDialogOpen: false,
       rightDialogOpen: false,
+      noticeDialogOpen: false,
     },
   },
   players: [