+++ src/views/NormalBattle/BattleAnimation/index.ts
@@ -1,0 +1,2 @@
+import BattleAnimation from "./BattleAnimation";
+export default BattleAnimation;