+++ src/views/NormalBattle/VS/VS.tsx
@@ -18,7 +18,6 @@ function VS() {
   const currentPlayer = players[currentPlayerIdx];
   const { job, gender, color } = currentPlayer;
   const battleCharacter = useBattleCharacter();
-  if (!battleCharacter) throw new Error("no battle character");
   const { characterType } = battleCharacter;
 
   const src1 = `${backendBaseUrl}/imgs/${job}_${gender}_${color}_vs.png`;